+++ wxc_sdk/as_api.py
@@ -19013,8 +19013,8 @@ class AsSCIM2GroupsApi(AsScimApiChild, base='identity/scim'):
         self,
         org_id: str,
         group_id: str,
-        schemas: list[str],
-        operations: list[PatchUserOperation],
+        schemas: list[str] = None,
+        operations: list[PatchUserOperation] = None,
     ) -> ScimGroup:
         """
         Update a group with PATCH
@@ -19130,6 +19130,9 @@ class AsSCIM2GroupsApi(AsScimApiChild, base='identity/scim'):
         :rtype: :class:`ScimGroup`
         """
         body = dict()
+        if operations is None:
+            raise ValueError('operations cannot be None')
+        schemas = schemas or ['urn:ietf:params:scim:api:messages:2.0:PatchOp']
         body['schemas'] = schemas
         body['Operations'] = TypeAdapter(list[PatchUserOperation]).dump_python(
             operations, mode='json', by_alias=True, exclude_none=True